Create more of a project structure in org.bitcoins.core.api, move DbRowAutoInc into the core project
