--- core/src/main/scala/org/bitcoins/core/api/NodeChainQueryApi.scala
@@ -1,3 +1,0 @@
-package org.bitcoins.core.api
-
-case class NodeChainQueryApi(nodeApi: NodeApi, chainQueryApi: ChainQueryApi)
