+++ core/src/main/scala/org/bitcoins/core/api/core/CoreApi.scala
@@ -1,4 +1,4 @@
-package org.bitcoins.core.api
+package org.bitcoins.core.api.core
 
 import org.bitcoins.core.protocol.transaction.Transaction
 import org.bitcoins.core.psbt.PSBT
