+++ core/src/main/scala/org/bitcoins/core/api/node/NodeChainQueryApi.scala
@@ -1,0 +1,5 @@
+package org.bitcoins.core.api.node
+
+import org.bitcoins.core.api.chain.ChainQueryApi
+
+case class NodeChainQueryApi(nodeApi: NodeApi, chainQueryApi: ChainQueryApi)
