+++ fee-provider/src/main/scala/org/bitcoins/feeprovider/ConstantFeeRateProvider.scala
@@ -1,6 +1,6 @@
 package org.bitcoins.feeprovider
 
-import org.bitcoins.core.api.FeeRateApi
+import org.bitcoins.core.api.feeprovider.FeeRateApi
 import org.bitcoins.core.wallet.fee.FeeUnit
 
 import scala.concurrent.Future
